Packages Styling
Amended the php for the plugin to use png's and changed the styles
associated with the images and text. Added css to display the packages
as per the visual supplied by Dan. Haven't been able to test on the
iPad yet so may have to do some extra work later on.