Disallow duplicate background images
One of the main rules of performance is to use as few bytes as possible
to get the job done. Along those lines, URLs are best used just once
inside of CSS. If you have more than one class that needs to use the
same background image, then it's better to have one class that uses the
URL and simply apply that class to the various elements where it is
needed.