Animation and Browser Detection
Added some css 3 transitions that will animate the blue and white title
block in on page load. Also added a filter to the functions.php that
detects the browser and adds the appropriate class to the body tag.
There will need to be more work done on this once it's decided what
will happen with the random text generator.